replace comrak with termimad
General markdown parsing is really not a good fit for the terminal. We
prefer to display the text with faithful spacing and just add bold,
italics, and maybe color. Headers are always a line on their own. Using
any markdown parsing library that converts to AST, just to render back
to some monospace text, is really inefficient *and* a huge complexity
burden.

Termimad is made to parse just basic decorators, and write back to
something like a terminal. This is perfect for our use case, we just
had to adapt it to ratatui widgets instead of writing escape codes
directly.