update/event logic can pick links from model, immutable view()
On update/event we just need to find a visible link. To avoid mutation
in render/view(), which will lead to subtle bugs with event/render
ordering, just rely on keeping the terminal height in the model. We can
infer if a widget is rendered or not by its height and the terminal height.