try out tree-sitter
Start with replacing split_headers_and_images regexes with tree-sitter
parsing.

Tree sitter gives the byte-range in the original source of the AST node.

If we didn't edit the source - we definitely want to strip "*" so that
it looks cool - it would have been a perfect match, and everything would
just have been "byte ranges" like for an editer. A kind of overlay, so
to speak.

Then we also can't just wrap the source text and then parse. The line
wrapping depends on the context, for example, in a blockquote the new
line must start with some offset. Same for lists etc.